updated taglib

+++ taglib/mpeg/mpegproperties.cpp
@@ -15,20 +15,20 @@
  *                                                                         *
  *   You should have received a copy of the GNU Lesser General Public      *
  *   License along with this library; if not, write to the Free Software   *
- *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
- *   USA                                                                   *
+ *   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA         *
+ *   02110-1301  USA                                                       *
  *                                                                         *
  *   Alternatively, this file is available under the Mozilla Public        *
  *   License Version 1.1.  You may obtain a copy of the License at         *
  *   http://www.mozilla.org/MPL/                                           *
  ***************************************************************************/
 
-#include <TagLib/tdebug.h>
-#include <TagLib/tstring.h>
+#include <tdebug.h>
+#include <tstring.h>
 
-#include <TagLib/mpegproperties.h>
-#include <TagLib/mpegfile.h>
-#include <TagLib/xingheader.h>
+#include "mpegproperties.h"
+#include "mpegfile.h"
+#include "xingheader.h"
 
 using namespace TagLib;
 
@@ -221,7 +221,7 @@ void MPEG::Properties::read()
       double length = timePerFrame * d->xingHeader->totalFrames();
 
       d->length = int(length);
-      d->bitrate = d->length > 0 ? d->xingHeader->totalSize() * 8 / length / 1000 : 0;
+      d->bitrate = d->length > 0 ? (int)(d->xingHeader->totalSize() * 8 / length / 1000) : 0;
   }
   else {
     // Since there was no valid Xing header found, we hope that we're in a constant
